changes to search models
