+++ api/endpoints/models/general_models.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from typing import Optional
+from typing import Optional, Union
 
 
 class Limits(BaseModel):
@@ -21,3 +21,15 @@ class GeneralInput(BaseModel):
     page: int = 0
     sort_method: str = "position"
     folio: str = ""
+
+
+class FullText(BaseModel):
+    text: str
+    highlightColor: int
+    
+
+class FullNames(BaseModel):
+    display_name: str
+    text_name: str
+    link1: Union[str, None] = None
+    link2: Union[str, None] = None
